+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Button`: Change `hydrogen` and `lithium` modifier styles.
+- Feature: `Button`: Add `active` prop that switches modifier to `lithium` when true.
+- Feature: `TagButton`: Deprecate `selected` prop, use new `active` instead.
+- Fix: `ButtonGroup`: Fix styles.
+- Fix: `TagButton`: Fix styles.
+- Fix: `CheckableButton`: Fix styles.
 - Fix: `TextInputWithUnit`: Improve styles.
 - Fix: `ContributionCard`: Improve styles.
 - Fix: `AvatarWithTextAndBadge`: Add `border` on the avatar image.